admin-analytics: refactor overview page to load gradually
